pyproject.toml
[tool.poetry]
name = "velib-en-mieux"
version = "0.1.0"
description = ""
authors = ["Thomas Legrand <legrand.thomas5@hotmail.fr>"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^0.25.3"
requests = "^2.22"
numpy = "^1.18"
plotly_express = "^0.4.1"
fbprophet = "^0.5"
fastapi = "^0.47.1"
uvicorn = "^0.11.2"
cachier = "^1.2.8"
pymongo = {extras = ["srv"], version = "^3.10.1"}
dnspython = "^1.16.0"
python-dotenv = "^0.10.5"

[tool.poetry.dev-dependencies]
pylint = "^2.4"
autopep8 = "^1.4"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
